requirements.txt
pyTelegramBotAPI
fpdf
python-dotenv